mktables: Can't generate simple casing tables
The simple as opposed to the full casing tables weren't able to be
generated correctly because I was passing the name of a table instead of
its object.  Normally these tables are generated at all, but I turned
generation of them on temporarily for development purposes, and found
this bug.  So there is no test case that can be generated for this fix.