mktables: Change format name for Decomposition.pl
This table which is used only by Normalize.pm has a special format that
is used in no other table.  In preparation for adding annotations to it,
make a special format signifier for it.  This affects only one line at
the top that indicates to the user the format of the table.

The table has comments in it to say it is volatile and no one should use
this table besides Normalize.pm, which currently doesn't look at that
signifier.