Fix x2p and utils dependencies for static builds
The x2p and utils .PL extractions use the Cwd module, which is not
available in miniperl in a static build.  (For a dynamic build, miniperl
only loads the pure perl portion of Cwd, but building the whole extension
ensures that the pure perl part is correctly built and available.  It might
be cleaner just to always require the full perl.)