Band-aid to handle long API symbol on VMS.
For the first time we now have a symbol longer than 31 characters
in the API, not just internally or generated in xsubs.  That means
we have to provide the shortened name in the linker options file
used to produce the shareable image, but the documentation on the
CRC used to produce the short name is incomplete.  For a more
general solution, we may have to root through the demangler
database or parse the output of CC/WARN=ENABLE=NAMESHORTENED.