regcomp.c: Get rid of compiler warning.
This patch should remove a compiler warning that is currently only
showing up in one compiler.  It declares a debug-only variable to be
volatile, so should silence the warning that it is getting clobbered.
Since this variable is only used for debugging purposes, when DEBUGGING
is defined, performance should not be an issue.