Correct tests for B::{cstring,perlstring,cchar}
Fix a Perl precedence error that prevented several test cases for cchar() from
being run. As a result, fix those test cases.

Fix an ASCII assumption in the tests common to cstring() and perlstring().
(chr 127 is " on EBCDIC, which is most definitely printable, so won't be
replaced with octal there.)