reginclass: Remove redundant test
The previous re-ordering of this function makes it clear that this test
doesn't do anything.  It is testing the charclass bitmap, but that was
already done in the re-ordered block from a previous commit, so if it
didn't succeed there, it won't succeed here.

In fact, trying to understand why this code was here was what led me to
figure out that it wasn't, and that things could be sped up by doing the
reordering.