Exclude more exported but non-existent functions.
pp_reach, pp_rvalues, and pp_transr are aliases, but regen/opcode.pl
generates declarations for aliases as if they were C functions in
their own right.  This is necessary for aliases that refer to mathoms
(as most of them do) but seems less right for new entries that will
never refer to an actual function of the same name.  For now, these
need to be explicitly excluded when building the linker options file
on VMS.  Failure to do so breaks the build with undefined symbol
warnings.