authorial perlipc edit
Looking through perlipc to make little fixes, it quickly became clear to me
that it had been patched with code by people with a Perl coding style, and
sometimes with an English language style, both very different to my own.

This made it seem like an old patchwork tattercloth that didn't fit
together very well.  It was confusing and detracted from the overall
message.  I've tried to fix all this.  I started with a recent git pull
and edited.  The result is something that is once again internally
self-consistent.